hairfit_server/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
slowapi
pillow
pytest
httpx

